Atualização nos endpoints para se relacionar com o h2

+++ src/main/java/br/pucpr/expedya/repository/CompanhiaAereaRepository.java
@@ -4,5 +4,6 @@ import br.pucpr.expedya.model.CompanhiaAerea;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface CompanhiaAereaRepository extends JpaRepository<CompanhiaAerea, Integer> {
+public interface CompanhiaAereaRepository extends JpaRepository<CompanhiaAerea, Long> {
 }
+
