+++ src/main/java/br/pucpr/expedya/repository/PassagemRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import br.pucpr.expedya.model.Passagem;
 
-public interface PassagemRepository extends JpaRepository<Passagem, Integer> {
+public interface PassagemRepository extends JpaRepository<Passagem, Long> {
 }
+
